\name{muUtil-package}
\alias{muUtil-package}
\alias{muUtil}
\docType{package}

\title{
muStat Utility Functions
}
\description{
This package contains a collection of utility functions for 
package muStat. 

For a complete list, use library(help="muUtil"). 
}
\details{

\tabular{ll}{
Package: \tab muUtil\cr
Type: \tab Package\cr
Version: \tab 1.6.0\cr
Date: \tab 2010-09-17\cr
License: \tab GPL (>= 2)
}

}
\author{
  Knut M. Wittkowski \email{kmw@rockefeller.edu}, 
  Tingting Song \email{ttsong@gmail.com}

Maintainer: Tingting Song \email{ttsong@gmail.com}
}

\examples{
x <- 5
NoOp(x)
# [1] 5

len(c(1:20))
# [1] 20
}
\keyword{ package }
