\name{NoOp}
\alias{NoOp}

\title{ No Operation }
\description{
  NoOp does nothing to the input.
}
\usage{
NoOp(x)
}

\arguments{
  \item{x}{ input}
}

\value{
  x is returned.
}

\author{ Knut M. Wittkowski \email{kmw@rockefeller.edu} }

\examples{
x <- 5
NoOp(x)
# [1] 5
}

\keyword{utilities}
