\name{NoWarn}
\alias{NoWarn}

\title{ Suppress Warnings }
\description{
  NoWarn executes x and temporarily disables the display of warnings.
}
\usage{
NoWarn(x)
}

\arguments{
  \item{x}{ input }
}

\value{
  invisible(NULL)
}

\author{ Knut M. Wittkowski \email{kmw@rockefeller.edu} }

\examples{
NoWarn(0)
}

\keyword{ utilities }
