
\name{ULPrint}
\alias{ULPrint}

\title{ Print Data }
\description{
  ...
}
\usage{
ULPrint(aData, nRows=10, nCols=10, Title="")
}

\arguments{
  \item{aData}{ }
  \item{nRows}{ }
  \item{nCols}{ }
  \item{Title}{ }
}

\value{

}

\author{ Knut M. Wittkowski \email{kmw@rockefeller.edu} }

\examples{

}

\keyword{ utilities }
