
\name{ULPrint}
\alias{ULPrint}

\title{ Print Data }
\description{
  Prints a data object according to the specified size.
}
\usage{
ULPrint(aData, nRows=10, nCols=10, Title="")
}

\arguments{
  \item{aData}{ data }
  \item{nRows}{ no. of rows to print }
  \item{nCols}{ no. of cols to print }
  \item{Title}{ title of output}
}

\value{

}

\author{ Knut M. Wittkowski \email{kmw@rockefeller.edu} }

\examples{

}

\keyword{ utilities }
