\name{muUtil-package}
\alias{muUtil-package}
\alias{muUtil}
\docType{package}

\title{
muStat Utility Functions
}
\description{
This package contains a collection of utility functions for 
package muStat. 

For a complete list, use library(help="muUtil"). 
}
\details{

\tabular{ll}{
Package: \tab muUtil\cr
Type: \tab Package\cr
Version: \tab 1.5.0\cr
Date: \tab 2008-02-07\cr
License: \tab GPL (>= 2)
}

}
\author{
  Knut M. Wittkowski \email{kmw@rockefeller.edu}, 
  Tingting Song \email{ttsong@gmail.com}

Maintainer: Tingting Song \email{ttsong@gmail.com}
}

\seealso{ 
  \code{\link[muStat:muStat-package]{muStat}}, 
  \code{\link[muS2RC:muS2RC-package]{muS2RC}}
}
\examples{
x <- 5
NoOp(x)
# [1] 5

len(c(1:20))
# [1] 20
}
\keyword{ package }
